fix(build): keep the cargo cache digest stable and claim each rule pair once
- Move `CARGO_NET_GIT_FETCH_WITH_CLI` past the cache step, whose digest hashes every `CARGO`-prefixed variable
- Add `PROSE_SETTLE_PAIRS`, claiming each pair by its earlier rule where the scopes partition the catalog
- Emit the claim per row from `gha:plan`, dropping a full sweep from 1560 pair slots to 861
- Read every single-rule run through `Pipeline::format`, dropping the lint pass and diagnostics `run` discarded
- Prune superseded `corpus-stage` entries after each bake, holding the family to one entry
- Watch `.mise/lib` in the baseline staleness check, so a changed record shape rebakes
- Guard the subset probe with the runaway watchdog, which reached only the corpus sweep
- Split the corpus-sweep surface into `common/sweep.rs`, carrying the watchdog, `Absorbing`, and `swept`
- Move `pipeline`'s test body into `pipeline/tests/`, one file per surface behind the shared sentinels
- Pin a module's leading docstring through the `shuffled` mutation, which moved it out of seat zero
- Count the files the subset probe cannot read, which it dropped silently
- Split `scope`, `shard`, and `claim` from their environment reads, adding seventeen cases over the parsers
- Test `runs_behind` ahead of the seat walk it short-circuits
- Drop the `--test corpus` build from `built_binary`, redundant with `rust:probe`'s own line
- Drop the `mimalloc` dev-dependency, already reaching the harnesses through the default `native` feature
- Announce a truncated mutation walk rather than leaving it to a short denominator
- Pin `format`'s suppression short-circuit, `format_span` over a suppressed source, and what `Source::clone` carries
