+++ .mise/tasks/rust/bake
@@ -1,4 +1,5 @@
 #!/usr/bin/env -S bash -euo pipefail
+#MISE alias       = "bake"
 #MISE description = "Bake the baseline stage of a corpus at every line length, and its mutation set, for a later head cycle"
 
 source "$MISE_PROJECT_ROOT/.mise/lib/corpus.bash"
