+++ CONTRIBUTING.md
@@ -135,7 +135,7 @@ Every corpus task defaults to the interpreter's own standard library and takes a
 
 `rust:mutations` mutates the corpus `rust:settle` sweeps and formats each variant with the unstable-output notice on, so a rewrite that settles over the corpus as written but not over a parseable edit of it still surfaces. Each mutation edits the module's concrete syntax tree, reordering top-level statements and class-body members, widening and narrowing identifiers, wrapping a sample of calls' arguments in redundant parentheses, injecting comments and suppression directives, and converting the line endings to CRLF, with every variant compiled before it lands so the report names a defect in *Prose* rather than one the mutation introduced. A second positional argument moves the pass off its sixty-second budget, and `PROSE_SETTLE_MUTATIONS` names the directory the set lives in, one `rust:bake` wrote or one the pass writes itself when the directory is absent. The pass formats that set in place, so a directory reused across runs carries the previous run's output rather than the mutated input, and the regeneration line the task prints rebuilds the input. A module the tree parser does not model, a template string among them, yields no variant, so read the denominator on every count before comparing two runs.
 
-The fixture tree is itself a corpus, and the same probe sweeps it on every `cargo test`, failing the suite on any subset that needs a second pass at any swept `code-line-length`, because a subset that settles at one budget can still edit its own output at another. `rust:subsets` is that probe aimed at a wider corpus, narrowed to the shipped default budget so the pointed sweep's wall clock holds, with `PROSE_SETTLE_RULES` keeping only the subsets that touch the rules it names, `PROSE_SETTLE_SHARD` taking one `k/n` share of the pairs, and `PROSE_SETTLE_VERIFY` folding every pair beside its chained run and failing on a divergence. Adding the reported case to the tree therefore turns the report into a permanent guard in the same stroke.
+The fixture tree is itself a corpus, and the same probe sweeps it on every `cargo test`, failing the suite on any subset that needs a second pass at any swept `code-line-length`, because a subset that settles at one budget can still edit its own output at another. `rust:subsets` is that probe aimed at a wider corpus, narrowed to the shipped default budget so the pointed sweep's wall clock holds, with `PROSE_SETTLE_RULES` keeping only the subsets that touch the rules it names, `PROSE_SETTLE_PAIRS` choosing whether a scoped run claims every pair touching its rules or only the pairs its own rules open, `PROSE_SETTLE_SHARD` taking one `k/n` share of the pairs, and `PROSE_SETTLE_VERIFY` folding every pair beside its chained run and failing on a divergence. A set of runs whose scopes partition the catalog takes the second claim, so each pair is probed once across the set rather than once per endpoint. Adding the reported case to the tree therefore turns the report into a permanent guard in the same stroke.
 
 The `🪻 Corpus` workflow runs the sweeps against the pinned interpreter's standard library on every pull request touching the crate, the mutation pass as its own `🎨 Mutations` job beside `⚓ Settle`, restoring the stage `rust:bake` writes on each `main` push and the mutation set it writes once per revision of the generator, gating on the settle, mutation, and subset verdicts, rendering the delta into its job's step summary as a report, and running the subset probe as one row per rule family named by the family's badge, each scoped to the rules the pull request touches under `crate/src/rules/` where nothing else the workflow watches changed and carrying the whole family otherwise.
