+++ crate/Cargo.toml
@@ -93,7 +93,6 @@ unicode-width       = "0.2"
 assert_cmd            = "2"
 indoc                 = "2"
 insta                 = { version = "1.39", features = ["glob", "yaml"] }
-mimalloc              = "0.1"
 proptest              = "1"
 rstest                = "0.26"
 ruff_python_formatter = { git = "https://github.com/astral-sh/ruff", tag = "0.15.22" }
